Timer and Screen Message
Tank:

Add: Function to Enable and Disable the Player Input (Inputs, Tick for Mouse and Show Mouse cursor) and
Setup function as false in the Start to block player inputs.

UMG:

Add: UMG Module.
Add: ScreenMessage WBP.
Add: ScreenMessage class and reparent the WBP.
Add: Function to set text of ScreenMessage.

GameMode:

Add: CreateWidget of ScreenMessage  and add this to player screen.
Add: Set ScreenMessage in coutdown and GameOver messages.
Call function on Tank to Enable Inputs when coutdown is 0